requirements.txt
crispy_forms
south
debug_toolbar
djangorestframework
